Move "common" package from pkg/apis to pkg
This change has been performed because operator-sdk
openapi and k8s generator was trying to generate elements
from this package and did not contain any information to do it,
making generations fail.
