+++ pkg/3scale/amp/component/redis.go
@@ -2,7 +2,7 @@ package component
 
 import (
 	"fmt"
-	"github.com/3scale/3scale-operator/pkg/apis/common"
+	"github.com/3scale/3scale-operator/pkg/common"
 	"github.com/3scale/3scale-operator/pkg/helper"
 
 	appsv1 "github.com/openshift/api/apps/v1"
