+++ pkg/3scale/amp/component/system_postgresql_image.go
@@ -2,7 +2,7 @@ package component
 
 import (
 	"fmt"
-	"github.com/3scale/3scale-operator/pkg/apis/common"
+	"github.com/3scale/3scale-operator/pkg/common"
 	"github.com/3scale/3scale-operator/pkg/helper"
 
 	imagev1 "github.com/openshift/api/image/v1"
